bugfix: Temporarily renamed the component to fix the error,
    Module not found: Error: Can't resolve '../components/navigation-menu.ts' in '/docs/embroider-css-modules/templates/application.hbs'
