+++ docs/embroider-css-modules/app/templates/application.hbs
@@ -4,7 +4,7 @@
   <header class={{this.styles.header}}>
     <NavigationNarrator @skipTo="#main-content" />
 
-    <NavigationMenu
+    <NavigationMenuNew
       @menuItems={{array
         (hash label="Home" route="index")
         (hash label="Dashboard" route="dashboard")
