+++ docs/sample-v2-addon/src/components/navigation-menu-new.ts
@@ -1,19 +1,19 @@
 import Component from '@glimmer/component';
 
-import styles from './navigation-menu.css';
+import styles from './navigation-menu-new.css';
 
 type MenuItem = {
   label: string;
   route: string;
 };
 
-interface NavigationMenuSignature {
+interface NavigationMenuNewSignature {
   Args: {
     menuItems: MenuItem[];
     name?: string;
   };
 }
 
-export default class NavigationMenuComponent extends Component<NavigationMenuSignature> {
+export default class NavigationMenuNewComponent extends Component<NavigationMenuNewSignature> {
   styles = styles;
 }
