+++ docs/sample-v2-addon/src/index.ts
@@ -1,2 +1,2 @@
-export { default as NavigationMenu } from './components/navigation-menu';
+export { default as NavigationMenuNew } from './components/navigation-menu-new';
 export { default as UiPageNew } from './components/ui/page-new';
