+++ docs/sample-v2-addon/src/template-registry.ts
@@ -1,7 +1,7 @@
-import type NavigationMenuComponent from './components/navigation-menu';
+import type NavigationMenuNewComponent from './components/navigation-menu-new';
 import type UiPageNewComponent from './components/ui/page-new';
 
 export default interface SampleV2AddonRegistry {
-  NavigationMenu: typeof NavigationMenuComponent;
+  NavigationMenuNew: typeof NavigationMenuNewComponent;
   'Ui::PageNew': typeof UiPageNewComponent;
 }
